+++ dotnet/src/UJMW.DynamicWcfHost.net48/Properties/AssemblyInfo.cs
@@ -15,6 +15,6 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: Guid("46adc8b8-2f14-4a31-87ca-95b71ee09ea9")]
 
-[assembly: AssemblyVersion("5.1.0")]
-[assembly: AssemblyInformationalVersion("5.1.0")]
-//DONT ADD: [assembly: AssemblyFileVersion("5.1.0")] !!!
+[assembly: AssemblyVersion("5.2.0")]
+[assembly: AssemblyInformationalVersion("5.2.0")]
+//DONT ADD: [assembly: AssemblyFileVersion("5.2.0")] !!!